[DCAS-46] -- Tweak to the way the background color order is applied to the nav items of a hero.
Sets the first item to be the striped secondary (green in this case)
Sets the second and subsequent even items to be the striped primary darkX (dark blue in this case)
Sets the third and subsequent odd items to be the striped primary (light blue in this case)
Sets the last item, usually fourth to be the striped accent warm darkX (orange in this case)

Special instances:
if only 1 item: it does not set the secondary or warm accent of the first and last item, it sets the primary dark.
If only 2 items: it does not set the last item as the warm accent/orange.
If only 3 items: it does not set the third item as the light blue, but as the warm accent/orange.

This setup follows the way the items were set in figma.